refactor(tests): split flagged-string engine tests into submodule
Moves the 6 `test_flagged_string_*` tests and the
`make_flagged_string_rule` helper from `src/evaluator/engine/tests/mod.rs`
into a new focused submodule `string_flags_dispatch_tests.rs`,
mirroring the layout the meta-type tests already use
(`meta_default_clear_indirect_tests.rs`, `meta_offset_tests.rs`,
`meta_use_tests.rs`).

Addresses CodeRabbit PR #288 review thread PRRT_kwDOP5Naes6E-aov,
which was deferred in the initial follow-up because the file-size
concern pre-dated PR #288. Since we are already iterating on the
follow-up, the file size guidance is best applied now: parent
`tests/mod.rs` drops from 2920 to 2786 lines.

Tests are unchanged -- only relocated. `super::*` reaches the
`MagicRule`, `Operator`, `OffsetSpec`, `TypeKind`, `Value`,
`EvaluationContext`, `EvaluationConfig`, `evaluate_rules`,
`evaluate_single_rule`, and `LibmagicError` imports through the parent
module, paralleling the existing meta-type submodules.
